+++ xworker_dataobject/src/main/java/xworker/dataObject/java/MapDataObjectActions.java
@@ -227,15 +227,15 @@ public class MapDataObjectActions {
                 if(pageInfo.getStart() > matchedDatas.size()){
                     pageInfo.setStart(matchedDatas.size());
                 }
-                int toIndex = pageInfo.getStart() + pageInfo.getLimit();
+                int toIndex = (int) (pageInfo.getStart() + pageInfo.getLimit());
                 if(toIndex > matchedDatas.size()){ 
                     toIndex = matchedDatas.size();
                 }
-                int startIndex = pageInfo.getStart();
+                int startIndex = (int) pageInfo.getStart();
                 if(startIndex < 0){
                     startIndex = 0;
                 }    
-                pageInfo.setDatas( matchedDatas.subList(pageInfo.getStart(), toIndex));
+                pageInfo.setDatas( matchedDatas.subList((int) pageInfo.getStart(), toIndex));
             }else{
                 pageInfo.setDatas(matchedDatas);
             }
